% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{append_tome_vector}
\alias{append_tome_vector}
\title{Generalized write for appending to individual vector objects to a tome file}
\usage{
append_tome_vector(vec, tome, target)
}
\arguments{
\item{vec}{The vector to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}
}
\description{
Generalized write for appending to individual vector objects to a tome file
}
