% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{available_tome_stats}
\alias{available_tome_stats}
\title{Returns available stats in a tome object}
\usage{
available_tome_stats(tome)
}
\arguments{
\item{tome}{the location of the tome file to read}
}
\value{
the name of available stats to read
}
\description{
Returns available stats in a tome object
}
