% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{cbind_tome_dgCMatrix}
\alias{cbind_tome_dgCMatrix}
\title{Add columns to a sparse matrix stored in .tome format}
\usage{
cbind_tome_dgCMatrix(mat, tome, target)
}
\arguments{
\item{mat}{The matrix to bind to an existing matrix}

\item{tome}{The .tome file to use}

\item{target}{The location in the .tome file to bind to}
}
\description{
Add columns to a sparse matrix stored in .tome format
}
