% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tome.R
\name{check_tome_data}
\alias{check_tome_data}
\title{Check the structure of data matrices stored in a .tome file}
\usage{
check_tome_data(tome, regions = "exon", orientations = "both",
  tome_ls = NULL, verbosity = NULL)
}
\arguments{
\item{verbosity}{optional: How verbose should the results be. Can be 0, 1, or 2. If NULL, uses global verbosity.}
}
\description{
Check the structure of data matrices stored in a .tome file
}
