% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tome.R
\name{check_tome_dgCMatrix_dims}
\alias{check_tome_dgCMatrix_dims}
\title{Check the structure of a sparse matrix stored in a .tome file}
\usage{
check_tome_dgCMatrix_dims(tome, target, dims, tome_ls = NULL,
  verbosity = NULL)
}
\arguments{
\item{tome}{the tome file to examine}

\item{target}{the location of the matrix within the tome file}

\item{dims}{the expected dimensions of the matrix (e.g. c(20,100) for 20 rows and 100 columns).}

\item{tome_ls}{optional: a pre-loaded list of tome contents generated by rhdf5::h5ls()}

\item{verbosity}{optional: How verbose should the results be. Can be 0, 1, or 2. If NULL, uses global verbosity.}
}
\value{
TRUE if all checks pass, errors will be generated if not.
}
\description{
Check the structure of a sparse matrix stored in a .tome file
}
