% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{check_tome_existence}
\alias{check_tome_existence}
\title{Check if an object in a tome file exists}
\usage{
check_tome_existence(tome, name)
}
\arguments{
\item{tome}{the tome file to check}

\item{name}{the name of the object in the tome file to check}
}
\value{
logical. TRUE if exists, FALSE if not.
}
\description{
Check if an object in a tome file exists
}
