% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{cl_to_anno}
\alias{cl_to_anno}
\title{Convert a cl factor object to an annotation data.frame}
\usage{
cl_to_anno(cl, cl.df = NULL, cl_col = "cl", base = "cluster")
}
\arguments{
\item{cl}{The cl factor object generated by clustering with scrattch.hicat}

\item{cl.df}{Optional: A cluster annotation data.frame.}

\item{cl_col}{The column to use to match cl.df to cl. Default is "cl", but if not found, will try to match using rownames(cl.df).}

\item{base}{The base of the cluster annotation. Default is "cluster", which will result in "cluster_id","cluster_label", and "cluster_color" columns.}
}
\value{
a data.frame with columns "sample_name", base"_id", base"_label", and base"_color"
}
\description{
Convert a cl factor object to an annotation data.frame
}
