% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{collapse_along}
\alias{collapse_along}
\title{Cumulatively collapse along a vector}
\usage{
collapse_along(x, collapse = "/")
}
\arguments{
\item{x}{The character vector to collapse}

\item{collapse}{The character to use for collapsing}
}
\description{
Cumulatively collapse along a vector
}
\examples{

x <- c("","data","exon","i")
collapse_along(x)

}
