% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_JSON.R
\name{dend_to_JSON}
\alias{dend_to_JSON}
\title{Convert a dendrogram into a nested JSON object}
\usage{
dend_to_JSON(dend)
}
\arguments{
\item{dend}{a dendrogram object}
}
\value{
a JSON text object
}
\description{
Convert a dendrogram into a nested JSON object
}
