% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_JSON.R
\name{dend_to_list}
\alias{dend_to_list}
\title{Convert a dendrogram object to a nested list}
\usage{
dend_to_list(dend)
}
\arguments{
\item{dend}{a dendrogram object}
}
\value{
a list object, as described above.
}
\description{
This function preserves node and leaf attributes as data.frame objects by converting each
nested dendrogram into a list containing the attribute data.frame and a nested list of children.
}
\details{
This provides a structure suitable for export to JSON.
}
