% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feather.R
\name{feather_to_list}
\alias{feather_to_list}
\title{Read a feather directory to a list of data.frames}
\usage{
feather_to_list(feather_dir = NULL, oldformat = F)
}
\arguments{
\item{feather_dir}{The directory containing .feather files.}

\item{oldformat}{Whether the new or old format are used}
}
\description{
This is likely to be deprecated soon.
}
