% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{flip_table}
\alias{flip_table}
\title{Transpose a gene x sample data.frame without losing a gene_name or sample_name column}
\usage{
flip_table(df, gene_col = "gene_name", sample_col = "sample_name")
}
\arguments{
\item{df}{The data.frame to transpose}

\item{gene_col}{The column used for gene names. Default = "gene_name".}

\item{sample_col}{The column used for sample names. Default = "sample_name".}
}
\description{
Transpose a gene x sample data.frame without losing a gene_name or sample_name column
}
