% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{group_annotations}
\alias{group_annotations}
\title{Group annotation columns}
\usage{
group_annotations(df, sample_col = "sample_name", keep_order = TRUE)
}
\arguments{
\item{df}{the annotation dataframe to arrange}

\item{sample_col}{the column with unique sample ids. Default is "sample_name".}

\item{keep_order}{a logical value. If FALSE, will sort the annotations alphanumerically by base.}
}
\value{
an annotation data frame with reordered columns
}
\description{
Group annotation columns
}
