% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{h5_delete}
\alias{h5_delete}
\title{Unlink an HDF5 object}
\usage{
h5_delete(h5_file, h5_object)
}
\arguments{
\item{h5_file}{The HDF5 file to edit.}

\item{h5_object}{The full name of the object within the HDF5 file to unlink.}
}
\description{
Unlink an HDF5 object
}
