% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{jagged_to_data.frame}
\alias{jagged_to_data.frame}
\title{Convert a jagged list of gene counts to a data.frame}
\usage{
jagged_to_data.frame(jagged, rows = c("sample_names", "gene_names"),
  cols = c("gene_names", "sample_names"))
}
\arguments{
\item{jagged}{The jagged list object to convert.}

\item{rows}{Character, either "sample_names" or "gene_names".}

\item{cols}{Character, either "sample_names" or "gene_names".}
}
\description{
Convert a jagged list of gene counts to a data.frame
}
