% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_matrix.R
\name{large_dgCMatrix_to_matrix}
\alias{large_dgCMatrix_to_matrix}
\title{Convert a large dgCMatrix object to a matrix}
\usage{
large_dgCMatrix_to_matrix(mat, chunk_size = 5000)
}
\arguments{
\item{mat}{The dgCMatrix object to convert}

\item{chunk_size}{The number of columns to read as a chunk. For ~30k genes, a chunk of 5000 (the default) is ~1GB in memory, and fits within the CHOLMOD limits.}
}
\value{
A standard matrix object in the same orientation as the dgCMatrix object.
}
\description{
May be useful in cases where calling as.matrix() directly gives a "Cholmod error 'problem too large'" error.
}
\details{
This works through the matrix in column-based chunks becaues this is much faster for conversion from column-indexed dgCMatrices.

It should still work on other types of Matrix-package classes, but may be slower.
}
