% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_matrix.R
\name{large_matrix_to_dgCMatrix}
\alias{large_matrix_to_dgCMatrix}
\title{Convert a large matrix object to a dgCMatrix}
\usage{
large_matrix_to_dgCMatrix(mat, chunk_size = 5000)
}
\arguments{
\item{mat}{The matrix file to convert}

\item{chunk_size}{The number of columns to read as a chunk. For ~30k genes, a chunk of 5000 (the default) is ~1GB in memory.}
}
\value{
A dgCMatrix object in the same orientation as the matrix object.
}
\description{
May be useful in cases where calling Matrix() directly gives a "negative length vectors are not allowed" error.
}
\details{
This works through the matrix in column-based chunks becaues this is much faster for conversion to column-indexed dgCMatrices.
}
