% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x.R
\name{read_10x_dgCMatrix}
\alias{read_10x_dgCMatrix}
\title{Read a whole sparse matrix directly from a 10X-style .h5 file}
\usage{
read_10x_dgCMatrix(h5, target)
}
\arguments{
\item{h5}{.h5 file to read.}

\item{target}{The sparse matrix to read within the .h5 file.}
}
\description{
Read a whole sparse matrix directly from a 10X-style .h5 file
}
