% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dttch.R
\name{read_dttch_anno}
\alias{read_dttch_anno}
\title{Read annotations table from a dttch file}
\usage{
read_dttch_anno(dttch_file, groups = NULL)
}
\arguments{
\item{dttch_file}{The location of the dttch file to read.}

\item{groups}{The groups to read - matches column names using grep. Can provide multiple with c(). If NULL, will get all columns. Default is NULL.}
}
\description{
Read annotations table from a dttch file
}
