% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dttch.R
\name{read_dttch_sample_data}
\alias{read_dttch_sample_data}
\title{Read Sample Expression Data from a dttch file}
\usage{
read_dttch_sample_data(dttch_file, samples, regions = "exon",
  type = "counts", transform = "none", form = "data.frame")
}
\arguments{
\item{dttch_file}{dttch file to read. Required.}

\item{samples}{A vector of sample names to read. Required}

\item{regions}{The gene regions to use. Can be "exon", "intron", or "both". Default = "exon".}

\item{transform}{Transformation to apply to values. Can be "none", "log", "log2", "log10". Log transforms will add 1 to values before transformation. Default = "none".}

\item{form}{The format of the output. Can be "data.frame", or "matrix". Default is "data.frame".}

\item{values}{The type of values to return. Can be "counts" or "cpm". Default = "counts".}
}
\value{
A data.frame with gene_name as the first column and each subsequent column
containing gene expression values and named for the samples; Or a matrix with columns as samples and rows as genes.
}
\description{
Read Sample Expression Data from a dttch file
}
