% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{read_gene_data_hdf5}
\alias{read_gene_data_hdf5}
\title{Read Gene Expression Data from an HDF5 file}
\usage{
read_gene_data_hdf5(hdf5_file, genes, type = c("counts", "cpm"))
}
\arguments{
\item{hdf5_file}{HDF5 file to read}

\item{genes}{A vector of gene names to read}

\item{type}{Whether to return "counts" or normalize to "cpm". Default is "counts".}
}
\value{
A data.frame with sample_id as the first column and each subsequent column
containing gene expression values and named for the genes.
}
\description{
Read Gene Expression Data from an HDF5 file
}
