% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad_dgCMatrix}
\alias{read_h5ad_dgCMatrix}
\title{Read a whole sparse matrix directly from a .h5ad file}
\usage{
read_h5ad_dgCMatrix(h5ad, target = "/raw.X")
}
\arguments{
\item{h5ad}{.h5ad file to read.}

\item{target}{The sparse matrix to read within the .h5ad file. Default = "/raw.X"}
}
\description{
Read a whole sparse matrix directly from a .h5ad file
}
