% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loom.R
\name{read_loom_projections}
\alias{read_loom_projections}
\title{Read Loom projections}
\usage{
read_loom_projections(loom_file)
}
\arguments{
\item{loom_file}{The loom file to read}
}
\value{
A data.frame with projection values as columns and samples as rows. The Loom CellID becomes the first column, sample_name.
Loom doesn't use a prefix to indicate paired coordinates, so you'll have to figure these out on your own.
}
\description{
Read Loom projections
}
