% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_data.frame}
\alias{read_tome_data.frame}
\title{Read a data.frame from a tome file}
\usage{
read_tome_data.frame(tome, df_name, stored_as = "data.frame",
  columns = NULL, match_type = "exact", get_all = FALSE)
}
\arguments{
\item{tome}{tome file to read}

\item{df_name}{character, the name of the data.frame object in the tome file structure}

\item{stored_as}{character, the storage mode of the data.frame used in write_tome_data.frame. Default = "data.frame".
If "data.frame", will read the df_name as a compound object. If "vectors" will read separate column vectors and build a data.frame.}

\item{columns}{character vector, specific columns to read (only if the object was stored as vectors).}

\item{get_all}{logical, whether or not to append all other columns after the specified columns (only if the object was stored as vectors).}
}
\description{
Read a data.frame from a tome file
}
