% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_data_dims}
\alias{read_tome_data_dims}
\title{Get dims for data stored in a tome file.}
\usage{
read_tome_data_dims(tome, transpose = F)
}
\arguments{
\item{tome}{tome file to read.}

\item{transpose}{logical indicating which orientation to get dims. Default = FALSE.
If FALSE, gives dimensions with samples as rows and genes as columns.
If TRUE,  gives dimensions for genes as rows and samples as columns.}
}
\description{
Get dims for data stored in a tome file.
}
