% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_dend}
\alias{read_tome_dend}
\title{Read dendrogram object from a tome file}
\usage{
read_tome_dend(tome, dend_name = NULL)
}
\arguments{
\item{tome}{the location of the tome file to read.}

\item{dend_name}{the dendrogram to read.}
}
\description{
Read dendrogram object from a tome file
}
