% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_dgCMatrix}
\alias{read_tome_dgCMatrix}
\title{Read a whole sparse matrix directly from a tome file}
\usage{
read_tome_dgCMatrix(tome, target)
}
\arguments{
\item{tome}{Tome file to read.}

\item{target}{The sparse matrix to read within the tome file.}
}
\description{
Read a whole sparse matrix directly from a tome file
}
