% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_gene_data}
\alias{read_tome_gene_data}
\title{Read Gene Expression Data from a tome file}
\usage{
read_tome_gene_data(tome, genes, regions = "exon", type = "counts",
  transform = "none", form = "data.frame")
}
\arguments{
\item{tome}{tome file to read. Required.}

\item{genes}{A vector of gene names to read. Required.}

\item{regions}{The gene regions to use. Can be "exon", "intron", or "both". Default = "exon".}

\item{transform}{Transformation to apply to values. Can be "none", "log", "log2", "log10". Log transforms will add 1 to values before transformation. Default = "none".}

\item{form}{The format of the output. Can be "data.frame", or "matrix". Default is "data.frame".}

\item{values}{The type of values to return. Can be "counts" or "cpm". Default = "counts".}
}
\value{
A data.frame with sample_name as the first column and each subsequent column
containing gene expression values and named for the genes; Or a matrix with columns as genes and rows as samples.
}
\description{
Read Gene Expression Data from a tome file
}
