% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_gene_meta}
\alias{read_tome_gene_meta}
\title{Read gene metadata table from a tome file}
\usage{
read_tome_gene_meta(tome, columns = NULL)
}
\arguments{
\item{tome}{The location of the tome file to read.}

\item{columns}{Specific columns to read If NULL, will get all columns. Default is NULL.}
}
\description{
Read gene metadata table from a tome file
}
