% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_gene_names}
\alias{read_tome_gene_names}
\title{Get all gene names in a tome file}
\usage{
read_tome_gene_names(tome)
}
\arguments{
\item{tome}{Tome file to read.}
}
\description{
Get all gene names in a tome file
}
