% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_genes_jagged}
\alias{read_tome_genes_jagged}
\title{Read tome gene count data as a jagged list}
\usage{
read_tome_genes_jagged(tome, genes, regions = "exon")
}
\arguments{
\item{tome}{the tome file to read.}

\item{genes}{a character vector of genes to read from the tome.}

\item{regions}{Which regions to retrieve. Can be "exon", "intron", or "both".}
}
\description{
Read tome gene count data as a jagged list
}
