% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_intron_lengths}
\alias{read_tome_intron_lengths}
\title{Get intron lengths from a tome file}
\usage{
read_tome_intron_lengths(tome, genes = NULL, return_as = "vector")
}
\arguments{
\item{tome}{tome file to read.}

\item{genes}{specific genes to retrieve}

\item{return_as}{If "vector", will return only the length values as a vector.
If "data.frame", will return a data.frame with two columns: "gene_name" and "intron_length".}
}
\description{
Get intron lengths from a tome file
}
