% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_projection}
\alias{read_tome_projection}
\title{Read projection coordinates table from a tome file}
\usage{
read_tome_projection(tome, proj_name = NULL)
}
\arguments{
\item{tome}{the location of the tome file to read.}

\item{proj_name}{the projection to read}
}
\description{
Read projection coordinates table from a tome file
}
