% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_sample_names}
\alias{read_tome_sample_names}
\title{Get all sample names in a tome file}
\usage{
read_tome_sample_names(tome)
}
\arguments{
\item{tome}{tome file to read.}
}
\description{
Get all sample names in a tome file
}
