% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_samples_jagged}
\alias{read_tome_samples_jagged}
\title{Read tome sample count data as a jagged list}
\usage{
read_tome_samples_jagged(tome, samples, regions = "exon")
}
\arguments{
\item{tome}{the tome file to read.}

\item{samples}{a character vector of genes to read from the tome.}

\item{regions}{Which regions to retrieve. Can be "exon", "intron", or "both".}
}
\description{
Read tome sample count data as a jagged list
}
