% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_serialized}
\alias{read_tome_serialized}
\title{Read a serialized object from a tome file}
\usage{
read_tome_serialized(tome, target)
}
\arguments{
\item{tome}{tome file to read}

\item{target}{character, the name of the serialized object in the tome file structure}
}
\description{
Read a serialized object from a tome file
}
