% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_stats}
\alias{read_tome_stats}
\title{Read stats table from a tome file}
\usage{
<<<<<<< HEAD
read_tome_stats(tome, stats_name, columns = NULL, get_all = FALSE)
=======
read_tome_stats(tome, stats_name = NULL, columns = NULL, get_all = FALSE)
>>>>>>> 671967780e5aa1cf6e6f548dca3a2d6d438a9809
}
\arguments{
\item{tome}{the location of the tome file to read.}

\item{stats_name}{the name of the stats table to read}

\item{columns}{selected columns to read. If NULL, reads all columns. Default = NULL.}

\item{get_all}{logical, whether or not to append all other columns after the specified columns. Default = FALSE.}
}
\description{
Read stats table from a tome file
}
