% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_stats_desc}
\alias{read_tome_stats_desc}
\title{Read stats descriptions table from a tome file}
\usage{
read_tome_stats_desc(tome)
}
\arguments{
\item{tome}{The location of the tome file to read.}
}
\description{
Read stats descriptions table from a tome file
}
