% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome.R
\name{read_tome_total_counts}
\alias{read_tome_total_counts}
\title{Get total per-gene counts from a tome-file.}
\usage{
read_tome_total_counts(tome, region = "exon")
}
\arguments{
\item{tome}{tome file to read.}

\item{region}{the gene regions to use. Can be "exon", "intron", or "both". Default = "exon"}
}
\description{
Useful for computing CPM or FPKM
}
