% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tome_lowlevel.R
\name{read_tome_vector}
\alias{read_tome_vector}
\title{Read a vector from a tome (or other HDF5 object)}
\usage{
read_tome_vector(tome, name)
}
\arguments{
\item{tome}{A tome file (other HDF5 files will work as well).}

\item{name}{The name of the object in the HDF5 hierarchy.}
}
\value{
A vector of the same type as the object in the HDF5 file.
}
\description{
The h5read function sticks close to the actual object stored in an HDF5 file - that is,
it will return a 1d array instead of a true vector object. This function is a simple
wrapper around h5read that will apply unlist and as.vector to the output of h5read so that
the result is a standard R vector.
}
