% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{save_sparse_matrix_h5}
\alias{save_sparse_matrix_h5}
\title{Save a dgCMatrix to HDF5 format}
\usage{
save_sparse_matrix_h5(mat, out_file, cols_are = "sample_id",
  overwrite = F, compression_level = 0)
}
\arguments{
\item{mat}{The matrix to be written}

\item{out_file}{The HDF5 file to write to}

\item{cols_are}{Specifies whether columns in the matrix are sample_ids or genes}

\item{overwrite}{Whether or not to overwrite an existing out_file}

\item{compression_level}{The data compression level for large HDF5 matrix objects}
}
\description{
Save a dgCMatrix to HDF5 format
}
