% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_scrattch.io_global_overwrite}
\alias{set_scrattch.io_global_overwrite}
\title{Set global overwrite parameter}
\usage{
set_scrattch.io_global_overwrite(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical, the setting of the default overwrite parameter for all scrattch.io functions.
Default = FALSE.}
}
\description{
This parameter will affect the default behavior of all write_tome functions, but will be overridden by parameters passed to each function.
}
