% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_scrattch.io_global_verbosity}
\alias{set_scrattch.io_global_verbosity}
\title{Set global verbosity parameter}
\usage{
set_scrattch.io_global_verbosity(verbosity = 2)
}
\arguments{
\item{verbosity}{integer, the verbosity level to use.

There are 3 levels:
0 removes all messages.
1 provides TRUE or FALSE for success or failure of writes. Reads are silent.
2 provides useful feedback for interactive use.
Default is 2.}
}
\description{
This parameter will affect the default behavior of all scrattch.io read or write functions that return print or cat statements
}
