% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{varibow}
\alias{varibow}
\title{Generate a rainbow palette with variation in saturation and value}
\usage{
varibow(n_colors)
}
\arguments{
\item{n_colors}{The number of colors to generate}
}
\description{
Generate a rainbow palette with variation in saturation and value
}
