% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv.R
\name{write_dgCMatrix_csv}
\alias{write_dgCMatrix_csv}
\title{Write a dgCMatrix to CSV in reasonable chunks}
\usage{
write_dgCMatrix_csv(mat, filename, col1_name = "gene", chunk_size = 1000)
}
\arguments{
\item{mat}{the dgCMatrix to write}

\item{filename}{the target CSV file}

\item{col1_name}{the name of the first column to write}

\item{chunk_size}{the number of rows to retrieve per chunk. Lower = less RAM usage.}
}
\description{
This function makes it so that we don't have to explode a dgCMatrix to a full matrix
in order to write the matrix to a CSV file with a named first column derived from rownames(mat).
}
