% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dttch.R
\name{write_dttch_anno}
\alias{write_dttch_anno}
\title{Write an annotations table to a dttch file.}
\usage{
write_dttch_anno(anno, out_file, mode = "add")
}
\arguments{
\item{anno}{The annotations data.frame to write. The first column must be "sample_id" or "sample_name".}

\item{out_file}{The dttch file to write to}

\item{mode}{The write mode. "replace" will remove existing annotations and replace it with anno. "add" will only add columns not already found. Default = "add".}
}
\description{
Write an annotations table to a dttch file.
}
