% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dttch.R
\name{write_dttch_data}
\alias{write_dttch_data}
\title{Save a both exon and intron counts to an HDF5 file (dttch)}
\usage{
write_dttch_data(exon_mat = NULL, intron_mat = NULL,
  out_file = "counts.dttch", cols_are = "sample_name", overwrite = F,
  compression_level = 4)
}
\arguments{
\item{exon_mat}{The exon matrix to store in dgCMatrix format}

\item{intron_mat}{The intron matrix to store in dgCMatrix format}

\item{out_file}{The HDF5 file to write to}

\item{cols_are}{Specifies whether columns in the matrix are sample_ids or genes}

\item{overwrite}{Whether or not to overwrite an existing out_file}

\item{compression_level}{The data compression level for large HDF5 matrix objects. default = 4.}
}
\description{
Save a both exon and intron counts to an HDF5 file (dttch)
}
