% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_sifter_stats_to_tome}
\alias{write_sifter_stats_to_tome}
\title{Caculate default stats for sifter and then write to tome}
\usage{
write_sifter_stats_to_tome(tome, overwrite = NULL)
}
\arguments{
\item{tome}{Path to the target tome file.}

\item{overwrite}{Whether or not to overwrite existing annotations. Default is NULL, which will use the global settings defined with set_scrattch.io_global_overwrite().}
}
\description{
Caculate default stats for sifter and then write to tome
}
