% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_anno_desc}
\alias{write_tome_anno_desc}
\title{Write annotation desc table to a tome file.}
\usage{
write_tome_anno_desc(anno_desc, tome)
}
\arguments{
\item{anno_desc}{The desc data.frame to write.}

\item{tome}{Path to the target tome file.}
}
\description{
Write annotation desc table to a tome file.
}
