% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/write_tome.R, R/write_tome_lowlevel.R
=======
% Please edit documentation in R/write_tome_lowlevel.R
>>>>>>> 671967780e5aa1cf6e6f548dca3a2d6d438a9809
\name{write_tome_data.frame}
\alias{write_tome_data.frame}
\alias{write_tome_data.frame}
\title{Generalized write for data.frames to a tome file}
\usage{
write_tome_data.frame(df, tome, target, store_as = "vectors",
  overwrite = NULL)
<<<<<<< HEAD

write_tome_data.frame(df, tome, target, store_as = "vectors",
  overwrite = NULL)
=======
>>>>>>> 671967780e5aa1cf6e6f548dca3a2d6d438a9809
}
\arguments{
\item{df}{The data.frame to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}

\item{store_as}{Either "data.frame", which will store as a compound object; or
"vectors", which stores each column as a separate object.}

\item{overwrite}{logical, whether or not to overwrite existing objects. Default = FALSE.}
<<<<<<< HEAD

\item{df}{The data.frame to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}

\item{store_as}{Either "data.frame", which will store as a compound object; or
"vectors", which stores each column as a separate object.}
=======
>>>>>>> 671967780e5aa1cf6e6f548dca3a2d6d438a9809
}
\description{
This function currently only works in an overwrite mode. Anything at the target
location will be removed, and the new df will be written.

This function currently only works in an overwrite mode. Anything at the target
location will be removed, and the new df will be written.
}
