% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{write_tome_dgCMatrix}
\alias{write_tome_dgCMatrix}
\title{Generalized write for dgCMatrix objects to a tome file}
\usage{
write_tome_dgCMatrix(mat, tome, target, overwrite = NULL,
  compression_level = 4)
}
\arguments{
\item{mat}{The dgCMatrix object to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}

\item{overwrite}{logical, whether or not to overwrite existing objects. Default = FALSE.}

\item{compression_level}{The compression level to use for long vectors x and i between 0 and 9. Default = 4.}
}
\description{
Generalized write for dgCMatrix objects to a tome file
}
