% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_exon_lengths}
\alias{write_tome_exon_lengths}
\title{Write exon lengths to a tome file.}
\usage{
write_tome_exon_lengths(exon_lengths, tome, overwrite = NULL)
}
\arguments{
\item{exon_lengths}{A data.frame containing the columns "gene_name" and "exon_length".}

\item{tome}{Path to the target tome file.}
}
\description{
Write exon lengths to a tome file.
}
