% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_gene_meta}
\alias{write_tome_gene_meta}
\title{Write gene info table to a tome file.}
\usage{
write_tome_gene_meta(genes, tome, overwrite = NULL)
}
\arguments{
\item{genes}{The genes data.frame to write.}

\item{tome}{Path to the target tome file.}

\item{overwrite}{Whether or not to overwrite existing annotations. Default is NULL, which will use the global settings defined with set_scrattch.io_global_overwrite().}
}
\description{
Write gene info table to a tome file.
}
