% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{write_tome_group}
\alias{write_tome_group}
\title{Generate a new group in a tome file}
\usage{
write_tome_group(tome, target_path)
}
\arguments{
\item{tome}{Path to the target tome file}

\item{target_path}{The target location within the tome file}
}
\description{
Generate a new group in a tome file
}
