% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_mapping}
\alias{write_tome_mapping}
\title{Write a mapping frequencies table to a tome file.}
\usage{
write_tome_mapping(mapping, mapping_name = NULL, tome, overwrite = NULL)
}
\arguments{
\item{mapping}{A data.frame with mapping results to write.}

\item{mapping_name}{The base name of the mapping Should match the mapping description table}

\item{tome}{Path to the target tome file.}
}
\description{
Write a mapping frequencies table to a tome file.
}
