% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_projection}
\alias{write_tome_projection}
\title{Write projection coordinates (e.g. tSNE or PCA) to a tome file.}
\usage{
write_tome_projection(proj, proj_name = NULL, tome, overwrite = NULL)
}
\arguments{
\item{proj}{A data.frame with projection coordinates to write.
Requires columns: sample_name, x, y (, z optional)}

\item{proj_name}{The base name of the projection. Should match the projection description table}

\item{tome}{Path to the target tome file.}

\item{overwrite}{Whether or not to overwrite existing annotations. Default is NULL, which will use the global settings defined with set_scrattch.io_global_overwrite().}
}
\description{
Write projection coordinates (e.g. tSNE or PCA) to a tome file.
}
