% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_projection_desc}
\alias{write_tome_projection_desc}
\title{Write a projection descriptions table to a tome file.}
\usage{
write_tome_projection_desc(proj_desc, tome, overwrite = NULL)
}
\arguments{
\item{proj_desc}{The desc data.frame to write.}

\item{tome}{Path to the target tome file.}
}
\description{
Write a projection descriptions table to a tome file.
}
