% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{write_tome_serialized}
\alias{write_tome_serialized}
\title{Generalized write for individual vector objects to a tome file}
\usage{
write_tome_serialized(obj, tome, target, overwrite = NULL)
}
\arguments{
\item{obj}{The R object to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}

\item{overwrite}{logical, whether or not to overwrite existing objects. Default = FALSE.}
}
\description{
Useful for R objects that can't easily be coerced to a data.frame or a vector, like lists or S3 classes.
Think of it like saveRDS() for HDF5 files.
}
