% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_tome_sifter_stats}
\alias{write_tome_sifter_stats}
\title{Caculate default stats for sifter and then write to tome}
\usage{
write_tome_sifter_stats(tome, overwrite = NULL)
}
\arguments{
\item{tome}{Path to the target tome file.1}

\item{overwrite}{Whether or not to overwrite existing annotations. Default is NULL, which will use the global settings defined with set_scrattch.io_global_overwrite().}
}
\description{
In this case, the target tome will need to have exon and intron data matrices, as well as annotations with the base "cluster".
}
