% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome_lowlevel.R
\name{write_tome_vector}
\alias{write_tome_vector}
\title{Generalized write for individual vector objects to a tome file}
\usage{
write_tome_vector(vec, tome, target, overwrite = FALSE)
}
\arguments{
\item{vec}{The vector to store}

\item{tome}{Path to the target tome file}

\item{target}{The target location within the tome file}

\item{overwrite}{logical, whether or not to overwrite existing objects. Default = FALSE.}
}
\description{
Generalized write for individual vector objects to a tome file
}
