% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{auto_annotate}
\alias{auto_annotate}
\title{Automatically format an annotation file}
\usage{
auto_annotate(anno, remove_factors = TRUE)
}
\arguments{
\item{anno}{an existing annotation data frame}

\item{remove_factors}{should factors be converted to character (currently, should be kept as TRUE)}
}
\value{
an updated data frame that has been automatically annotated properly
}
\description{
This takes an anno file as input at any stage and properly annotates it for compatability with
  shiny and other scrattch functions.  In particular, it ensures that columns have a label,
  an id, and a color, and that there are no factors.  It won't overwrite columns that have
  already been properly process.
}
