% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert_tome_to_feather}
\alias{convert_tome_to_feather}
\title{Expands a tome file to the expected feather and rda files for molgen-shiny}
\usage{
convert_tome_to_feather(tome, output_folder = getwd(),
  regions = "both", dend_name = "cluster")
}
\arguments{
\item{tome}{Path to the target tome file.1}

\item{output_folder}{Folder where output files should be written (default is current directory)}

\item{regions}{Which gene regions to use. Can be "both" (default"), "exon", or "intron".}

\item{dend_name}{Name of dendrogram to read (default="cluster")}
}
\value{
No R objects returned.
}
\description{
This file takes a tome as input and writes all of the necessary files for molgen-shiny to work properly in a feather file
}
