% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{create_desc}
\alias{create_desc}
\title{Create a generic description file}
\usage{
create_desc(dat, name = colnames(dat), use_label_columns = FALSE)
}
\arguments{
\item{dat}{any data frame that you would like to create a description file for}

\item{use_label_columns}{should only columns containing "_label" be included (default = FALSE)}

\item{names}{desired names of each element in the description file (default is the column names)}
}
\value{
a data.frame with columns "base", "name", and "type" for writing to tome
}
\description{
Create a generic description file
}
