% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_anno}
\alias{write_tome_anno}
\title{Write an annotations table to a tome file.}
\usage{
write_tome_anno(anno, tome, mode = "add")
}
\arguments{
\item{anno}{The annotations data.frame to write. The first column must be "sample_id" or "sample_name".}

\item{tome}{The tome file to write to}

\item{mode}{The write mode. "replace" will remove existing annotations and replace it with anno. "add" will only add columns not already found. Default = "add".}
}
\description{
Write an annotations table to a tome file.
}
