% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_data}
\alias{write_tome_data}
\title{Save a both exon and intron counts to an HDF5 file (tome)}
\usage{
write_tome_data(exon_mat = NULL, intron_mat = NULL, tome = "counts.tome",
  cols_are = "sample_name", overwrite = F, compression_level = 4)
}
\arguments{
\item{exon_mat}{The exon matrix to store in dgCMatrix format}

\item{intron_mat}{The intron matrix to store in dgCMatrix format}

\item{tome}{The HDF5 file to write to}

\item{cols_are}{Specifies whether columns in the matrix are sample_ids or genes}

\item{overwrite}{Whether or not to overwrite an existing tome}

\item{compression_level}{The data compression level for large HDF5 matrix objects. default = 4.}
}
\description{
Save a both exon and intron counts to an HDF5 file (tome)
}
