% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tome.R
\name{write_tome_stats}
\alias{write_tome_stats}
\title{Write a stats table (e.g. median expression per cluster) to a tome file.}
\usage{
write_tome_stats(stats, stats_name = NULL, tome, overwrite = FALSE)
}
\arguments{
\item{stats}{The stats data.frame to write.}

\item{tome}{Path to the target tome file.}
}
\description{
Write a stats table (e.g. median expression per cluster) to a tome file.
}
